Make the events module optional feature
