+++ src/event/sys/unix.rs
@@ -1,5 +1,7 @@
 #[cfg(feature = "event-stream")]
+#[cfg(feature = "events")]
 pub(crate) mod waker;
 
 pub(crate) mod file_descriptor;
+#[cfg(feature = "events")]
 pub(crate) mod parse;
