+++ src/terminal.rs
@@ -98,6 +98,7 @@ use crate::{csi, impl_display, Result};
 
 pub(crate) mod sys;
 
+#[cfg(feature = "events")]
 pub use sys::supports_keyboard_enhancement;
 
 /// Tells whether the raw mode is enabled.
