+++ src/terminal/sys.rs
@@ -1,10 +1,12 @@
 //! This module provides platform related functions.
 
 #[cfg(unix)]
+#[cfg(feature = "event")]
 pub use self::unix::supports_keyboard_enhancement;
 #[cfg(unix)]
 pub(crate) use self::unix::{disable_raw_mode, enable_raw_mode, is_raw_mode_enabled, size};
 #[cfg(windows)]
+#[cfg(feature = "event")]
 pub use self::windows::supports_keyboard_enhancement;
 #[cfg(windows)]
 pub(crate) use self::windows::{
